bot/requirements.txt
aiogram==3.1.1
asyncpg==0.29.0
pydantic>=2.1.1,<2.4
python-dotenv==1.0.0
aiohttp==3.8.6
